worker: splits the config-specific definitions to separate file.
